Interface GoodDao + UserDao
