+++ src/main/java/servlet/admin/DeleteGoodServlet.java
@@ -1,6 +1,7 @@
 package servlet.admin;
 
 import dao.GoodDao;
+
 import dao.GoodDaoHib;
 import org.apache.log4j.Logger;
 
@@ -14,14 +15,13 @@ import java.io.IOException;
 @WebServlet("/GoodServlet/delete")
 public class DeleteGoodServlet extends HttpServlet {
 	final static Logger logger = Logger.getLogger(DeleteGoodServlet.class);
-	private GoodDao goodDao = new GoodDao();
-	private GoodDaoHib goodDaoHib = new GoodDaoHib();
+	private GoodDao goodDao = new GoodDaoHib();
 
 	@Override
 	protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
 		logger.debug("good deleting");
 		int id = Integer.parseInt(req.getParameter("id"));
-		goodDaoHib.delete(goodDaoHib.findById(id));
+		goodDao.deleteGood(id);
 		resp.sendRedirect("/GoodServlet");
 	}
